Well, it still has issues, but at least it won't go killing off your X server any more....
git-svn-id: http://svn.enlightenment.org/svn/e/trunk@1193 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33

+++ e16/epplets/epplets/E-Xss.c
@@ -49,7 +49,9 @@ start_hack(void) {
 
 static void
 stop_hack(void) {
-  kill(hack_pid, SIGTERM);
+  if (hack_pid != -1) {
+    kill(hack_pid, SIGTERM);
+  }
   hack_pid = -1;
 }
